Standardize the definition of a UFS dinode
Each program that operates on UFS on-disk inodes defines its own
version of a dinode. They all (of necessity) define the same
layout but use different names. This change adds a definition of
a dinode (a union of a UFS1 on-disk inode and a UFS2 on-disk inode)
in sys/ufs/ufs/dinode.h.

This is based on Kirk McKusick's commit
aa90fbed151de512ab6e59f75df009533a15751f in FreeBSD.

Unfortunately "union dinode_u" can't be standardize because it also
includes ext2fs_dinode.